fix compile warning

+++ src/thttpbuffer.cpp
@@ -98,7 +98,7 @@ void THttpBuffer::parse()
             }
 
             lengthToRead = qMax(idx + 4 + (qint64)header.contentLength() - httpBuffer.length(), 0LL);
-            tSystemDebug("lengthToRead: %lld", lengthToRead);
+            tSystemDebug("lengthToRead: %d", (int)lengthToRead);
         }
     } else {
         tSystemWarn("Unreachable code in normal communication");
